Update README.md (#37)
Okay so this is a bit of a mess. I put this repository on READ-ONLY
mode, but learned from @isaacwasserman that this project is actually
being used by people and that Snazzy Labs on YouTube
(https://youtu.be/D2_8qJi2jpQ?si=YZPHdPG8TAt5lJH4&t=815) is a happy
user. Also I didn't know that this project is packaged in brew already.

I'll pick up maintenance again on this tomorrow.

Thanks @isaacwasserman for letting me know!